+++ toolchain/scripts/stage4/install_cereal.sh
@@ -42,20 +42,17 @@ case "$with_cereal" in
         # download from github.com and checksum
             echo "===> Notice: This version of CEREAL is downloaded in GitHub Release, which will always be out-of-date version <==="
             download_pkg_from_url "${cereal_sha256}" "${filename}" "${url}"
-            # echo "wget $url -O $filename"
-            # if ! wget $url -O $filename; then
-            # report_error "failed to download $url"
-            # recommend_offline_installation $filename $url
-            # fi
-            # # checksum
-            # checksum "$filename" "$cereal_sha256"
         fi
+    if [ "${pack_run}" = "__TRUE__" ]; then
+      echo "--pack-run mode specified, skip installation"
+    else
         echo "Installing from scratch into ${pkg_install_dir}"
         [ -d $dirname ] && rm -rf $dirname
         tar -xzf $filename
         mkdir -p "${pkg_install_dir}"
         cp -r $dirname/* "${pkg_install_dir}/"
         write_checksums "${install_lock_file}" "${SCRIPT_DIR}/stage4/$(basename ${SCRIPT_NAME})"
+    fi
     fi
         ;;
     __SYSTEM__)
